analyze.dart에서 빨간줄 삭제~
